iox-#1103 Refactor discovery wait example
